Restructure attTrackingErrorAlg Types.h to pure C
